Fix: moved "ValueObject" to Domain root
