--- AnimeAPI.Domain/Entities/ValueObjects/Exceptions/InvalidSizeNameException.cs
@@ -1,5 +1,0 @@
-using AnimeAPI.Domain.Shared.Exceptions;
-
-namespace AnimeAPI.Domain.Entities.ValueObjects.Exceptions;
-
-public class InvalidSizeNameException(string message) : DomainException(message);
