+++ AnimeAPI.Domain/ValueObjects/Exceptions/NameExceptions/InvalidNameSizeException.cs
@@ -1,0 +1,5 @@
+using AnimeAPI.Domain.Shared.Exceptions;
+
+namespace AnimeAPI.Domain.ValueObjects.Exceptions.NameExceptions;
+
+public class InvalidNameSizeException(string message) : DomainException(message);
