+++ AnimeAPI.Domain/ValueObjects/Exceptions/NameExceptions/NullNameException.cs
@@ -1,5 +1,5 @@
 using AnimeAPI.Domain.Shared.Exceptions;
 
-namespace AnimeAPI.Domain.Entities.ValueObjects.Exceptions;
+namespace AnimeAPI.Domain.ValueObjects.Exceptions.NameExceptions;
 
 public class NullNameException(string message) : DomainException(message);
